Maximum number of bytes that can be read in a single operation

+++ include/zenoh-pico/link/link.h
@@ -67,7 +67,7 @@ _zn_link_p_result_t _zn_open_link(const z_str_t locator, const clock_t tout);
 _zn_link_p_result_t _zn_listen_link(const z_str_t locator, const clock_t tout);
 
 int _zn_link_send_wbuf(const _zn_link_t *link, const _z_wbuf_t *wbf);
-int _zn_link_recv_zbuf(const _zn_link_t *link, _z_zbuf_t *zbf, z_bytes_t *addr);
-int _zn_link_recv_exact_zbuf(const _zn_link_t *link, _z_zbuf_t *zbf, size_t len, z_bytes_t *addr);
+size_t _zn_link_recv_zbuf(const _zn_link_t *link, _z_zbuf_t *zbf, z_bytes_t *addr);
+size_t _zn_link_recv_exact_zbuf(const _zn_link_t *link, _z_zbuf_t *zbf, size_t len, z_bytes_t *addr);
 
 #endif /* ZENOH_PICO_LINK_H */
